Use npm slick and bootstrap and autogenerate js and css
