+++ hugo/layouts/partials/footer.html
@@ -116,8 +116,4 @@ drift.load('9r2gag6k94nb');
 </script>
 <!-- End of Async Drift Code -->
 
-<!-- jQuery first, then Bootstrap JS. -->
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.4/jquery.min.js"></script>
-<script src="js-load/bootstrap.bundle.min.js"></script>
-<script type="text/javascript" src="js-load/slick.min.js"></script>
-<script type="text/javascript" src="js-load/script.js"></script>
+<script type="text/javascript" src="js/bundle.js"></script>
