+++ hugo/layouts/partials/head.html
@@ -20,17 +20,8 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
     <meta name="viewport" content="width=device-width, user-scalable=yes">
 
-    <link rel="stylesheet" href="css-load/bootstrap.min.css?v={{ now.Format "1136239445" }}">
     <link rel="stylesheet" href="icomoon/icons.css?v={{ now.Format "1136239445" }}">
-    <link rel="stylesheet" href="css-load/slick.css?v={{ now.Format "1136239445" }}">
-    <link rel="stylesheet" href="scss/style.css?v={{ now.Format "1136239445" }}">
-
-    <!-- <link href="//fonts.googleapis.com/css?family=Source+Sans+Pro:400,700|Roboto:400,700|Caveat:400" rel="stylesheet">
-    <link rel="stylesheet" href="//cdn.jsdelivr.net/font-hack/2.020/css/hack.min.css">
-    <link rel="stylesheet" type="text/css" href="//cdn.jsdelivr.net/jquery.slick/1.6.0/slick.css" />
-    <link rel="stylesheet" type="text/css" href="//cdn.jsdelivr.net/jquery.slick/1.6.0/slick-theme.css" />
-
-    <link href="css/style.css?v={{ now.Format "1136239445" }}" rel="stylesheet" type="text/css"> -->
+    <link rel="stylesheet" href="css/bundle.css?v={{ now.Format "1136239445" }}">
 
     <script>
     (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
