+++ src/sass/app.scss
@@ -4,3 +4,4 @@
 @import 'partials/layout';
 @import 'partials/cards';
 @import 'partials/google-calendar';
+@import 'partials/carousel';
