+++ src/sass/partials/_carousel.scss
@@ -1,0 +1,21 @@
+.slick-prev,
+.slick-next {
+    &:before{
+        content: '';
+        width: 32px;
+        height: 32px;
+        display: inline-block;
+        opacity: 1;
+        margin-top: -10px;
+      }
+}
+
+.slick-prev:before {
+    background: url(/../img/icons/carousel-previous.svg);
+    margin-left: -32px;
+  }
+
+.slick-next:before {
+    background: url(/../img/icons/carousel-next.svg);
+    margin-left: 10px;
+}
