+++ webpack.config.js
@@ -10,7 +10,8 @@ const IS_PRODUCTION = process.env.npm_lifecycle_event === 'build';
 
 const baseConf = {
   entry: [
-    path.join(__dirname, 'src/js/index.js')
+    path.join(__dirname, 'src/js/index.js'),
+    path.join(__dirname, 'src/sass/app.scss')
   ],
   output: {
     path: path.join(__dirname, 'static'),
@@ -60,7 +61,7 @@ const productionConf = {
         NODE_ENV: JSON.stringify('production')
       }
     }),
-    new ExtractTextPlugin({ filename: 'css/style.css', allChunks: true }),
+    new ExtractTextPlugin({ filename: 'css/bundle.css', allChunks: true }),
     new webpack.optimize.UglifyJsPlugin({
       compress: {
         warnings: false
@@ -75,7 +76,7 @@ const developmentConf = {
     contentBase: 'static/'
   },
   plugins: [
-    new ExtractTextPlugin({ filename: 'css/style.css', allChunks: true })
+    new ExtractTextPlugin({ filename: 'css/bundle.css', allChunks: true })
   ]
 };
